+++ yde-landlords/src/main/java/com/xt/landlords/game/mission/phase/MissionClearPhaseData.java
@@ -2,6 +2,8 @@ package com.xt.landlords.game.mission.phase;
 
 import org.sunyata.octopus.model.PhaseData;
 
+import java.math.BigDecimal;
+
 /**
  * Created by leo on 17/5/16.
  */
@@ -9,16 +11,16 @@ public class MissionClearPhaseData extends PhaseData {
 
     private String serialNo;
 
-    public float getTotalMoney() {
+    public BigDecimal getTotalMoney() {
         return totalMoney;
     }
 
-    public MissionClearPhaseData setTotalMoney(float totalMoney) {
+    public MissionClearPhaseData setTotalMoney(BigDecimal totalMoney) {
         this.totalMoney = totalMoney;
         return this;
     }
 
-    private float totalMoney;//获得的总奖金
+    private BigDecimal totalMoney;//获得的总奖金
 
     public MissionClearPhaseData setSerialNo(String serialNo) {
         this.serialNo = serialNo;
